+++ src/main/java/com/fatec/backend/configuration/UserAuthenticationFilter.java
@@ -1,9 +1,11 @@
 package com.fatec.backend.configuration;
 
 
-import com.fatec.backend.model.UserDetailsImpl;
+
+import com.fatec.backend.model.User.User;
+import com.fatec.backend.model.User.UserDetailsImpl;
 import com.fatec.backend.repository.UserRepository;
-import com.fatec.backend.model.User;
+
 import com.fatec.backend.service.auth.JwtTokenService;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
