+++ src/main/java/com/fatec/backend/mapper/vehicle/VehicleMapper.java
@@ -1,7 +1,7 @@
 package com.fatec.backend.mapper.vehicle;
 
 import com.fatec.backend.DTO.vehicle.VehicleDTO;
-import com.fatec.backend.model.Vehicle;
+import com.fatec.backend.model.vehicle.Vehicle;
 import org.mapstruct.Mapper;
 import org.mapstruct.factory.Mappers;
