+++ src/main/java/com/fatec/backend/repository/VehicleRespository.java
@@ -1,6 +1,6 @@
 package com.fatec.backend.repository;
 
-import com.fatec.backend.model.Vehicle;
+import com.fatec.backend.model.vehicle.Vehicle;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
