+++ src/main/java/com/fatec/backend/service/user/UserDetailsServiceImpl.java
@@ -1,7 +1,7 @@
 package com.fatec.backend.service.user;
 
-import com.fatec.backend.model.User;
-import com.fatec.backend.model.UserDetailsImpl;
+import com.fatec.backend.model.User.User;
+import com.fatec.backend.model.User.UserDetailsImpl;
 import com.fatec.backend.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UserDetails;
