+++ src/main/java/com/fatec/backend/service/user/UserService.java
@@ -6,16 +6,13 @@ import com.fatec.backend.DTO.auth.JwtTokenDTO;
 import com.fatec.backend.DTO.auth.LoginUserDTO;
 import com.fatec.backend.DTO.user.UpdateUserDTO;
 import com.fatec.backend.DTO.user.UserDTO;
-import com.fatec.backend.DTO.vehicle.VehicleDTO;
 import com.fatec.backend.exception.InvalidCredentialsException;
 import com.fatec.backend.exception.UserNotFoundException;
 import com.fatec.backend.mapper.user.UserMapper;
-import com.fatec.backend.mapper.vehicle.VehicleMapper;
-import com.fatec.backend.model.UserDetailsImpl;
-import com.fatec.backend.model.Vehicle;
+import com.fatec.backend.model.User.UserDetailsImpl;
 import com.fatec.backend.repository.UserRepository;
 import com.fatec.backend.configuration.SecurityConfig;
-import com.fatec.backend.model.User;
+import com.fatec.backend.model.User.User;
 import com.fatec.backend.service.auth.JwtTokenService;
 import lombok.AllArgsConstructor;
 import org.springframework.data.domain.Page;
